+++ pxr/usd/lib/pcp/types.h
@@ -166,7 +166,7 @@ struct Pcp_SdSiteRef : boost::totally_ordered<Pcp_SdSiteRef> {
 // Internal type for Sd sites.
 struct Pcp_CompressedSdSite {
     Pcp_CompressedSdSite(size_t nodeIndex_, size_t layerIndex_) :
-        nodeIndex(nodeIndex_), layerIndex(layerIndex_)
+        nodeIndex(static_cast<uint16_t>(nodeIndex_)), layerIndex(static_cast<uint16_t>(layerIndex_))
     {
         TF_VERIFY(nodeIndex_  < (1 << 16));
         TF_VERIFY(layerIndex_ < (1 << 16));
